Moved roles.py to common student
