+++ lms/djangoapps/bulk_email/tasks.py
@@ -37,7 +37,7 @@ from bulk_email.models import (
     SEND_TO_MYSELF, SEND_TO_ALL, TO_OPTIONS,
 )
 from courseware.courses import get_course, course_image_url
-from courseware.roles import CourseStaffRole, CourseInstructorRole
+from student.roles import CourseStaffRole, CourseInstructorRole
 from instructor_task.models import InstructorTask
 from instructor_task.subtasks import (
     SubtaskStatus,
