+++ lms/djangoapps/dashboard/tests/test_sysadmin.py
@@ -17,7 +17,7 @@ from django.utils.html import escape
 from django.utils.translation import ugettext as _
 import mongoengine
 
-from courseware.roles import CourseStaffRole, GlobalStaff
+from student.roles import CourseStaffRole, GlobalStaff
 from courseware.tests.tests import TEST_DATA_MONGO_MODULESTORE
 from dashboard.models import CourseImportLog
 from dashboard.sysadmin import Users
